# Back/docs/api/OpenAPI.yaml
openapi: 3.1.0
info:
  title: API de Gestion de Solicitudes Academicas
  version: 1.1.0
  description: >
    API REST domain-first para gestionar solicitudes academicas.
    El contrato esta alineado con los DTOs y commands de la capa de aplicacion
    del proyecto `co.edu.uniquindio.Proyecto_Avanzada`.
servers:
  - url: http://localhost:8080
    description: Entorno local
...
paths:
  /api/usuarios/registro:
    post:
      tags: [Usuarios]
      summary: Registrar usuario
      description: >
        Registra un nuevo usuario en el sistema con su informacion basica,
        estado de activacion y rol asociado.
      operationId: registrarUsuario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearUsuarioRequest'
      responses:
        '201':
          description: Usuario registrado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioResumenDTO'
        '400':
          description: Solicitud invalida o error de validacion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autenticado o token invalido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operacion no autorizada para el rol actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Identificacion:
      type: string
      minLength: 5
      maxLength: 20
      examples: ["1090123456"]
    Rol:
      type: string
      enum: [ESTUDIANTE, ADMINISTRATIVO, COORDINADOR, DOCENTE]
      examples: [ESTUDIANTE, ADMINISTRATIVO, COORDINADOR, DOCENTE]
...
    CrearUsuarioRequest:
      type: object
      additionalProperties: false
      required: [nombre, identificacion, correo, activo, rol]
      properties:
        nombre:
          type: string
          minLength: 3
          maxLength: 150
        identificacion:
          $ref: '#/components/schemas/Identificacion'
        correo:
          type: string
          format: email
          maxLength: 200
        activo:
          type: boolean
        rol:
          $ref: '#/components/schemas/Rol'
      examples: [{nombre: "Laura Fernanda Gomez", identificacion: "1090123456", correo: "laura.gomez@uqvirtual.edu.co", activo: true, rol: "ESTUDIANTE"}, {nombre: "carlos alonso barahona alvarz", identificacion:"1094896340", correo: "carlosa.barahonaa@uqvirutual.edu.co", activo: true, rol: "ESTUDIANTE"}]
...
    UsuarioResumenDTO:
      type: object
      additionalProperties: false
      required: [nombre, identificacion, correo, activo, rol]
      properties:
        nombre:
          type: string
          maxLength: 150
        identificacion:
          $ref: '#/components/schemas/Identificacion'
        correo:
          type: string
          maxLength: 200
        activo:
          type: boolean
        rol:
          type: string
          maxLength: 50
      examples: [{"nombre":"Laura Fernanda Gomez","identificacion":"1090123456","correo":"laura.gomez@uqvirtual.edu.co","activo":true,"rol":"ESTUDIANTE"}]
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
          maxLength: 100
        message:
          type: string
          maxLength: 500
        path:
          type: string
          maxLength: 300
        validationErrors:
          oneOf:
            - type: object
              additionalProperties:
                type: string
            - type: 'null'
      examples:
        [{"timestamp": "2026-04-06T11:25:00Z",
        "status": 400,
        "error": "Bad Request",
        "message": "La solicitud contiene datos invalidos.",
        "path": "/api/solicitudes",
        "validationErrors":
          {"descripcion":"La descripcion debe tener al menos 10 caracteres."}}]